Implemented Hodge diamonds
Hence I should change the name of the repository.

This is a quick and dirty port of Nicolas Addington's C++ code. There
are some overflow issues (but they don't seem to happen for the numbers
that one can input in the current input method).

And there still has to be a way to display the diamonds, bien sur.